% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccdata1a_doc.r
\docType{data}
\name{iccdata1a}
\alias{iccdata1a}
\title{Dataset of ratings assigned by 4 judges to 6 targets/subjects.}
\format{
This dataset has 6 rows (for the 6 subjects) and 5 columns. All 5
columns are mandatory when using ICC functions from this package.
 \describe{
    \item{target}{This variable represents the target or subject number}
    \item{J1}{All ratings from judge 1}
    \item{J2}{All ratings from judge 2}
    \item{J3}{All ratings from judge 3}
    \item{J4}{All ratings from judge 4}
 }
}
\source{
Gwet, K.L. (2021) \emph{Handbook of Inter-Rater Reliability:
Volume 1}, 5th Edition. AgreeStat Analytics.
}
\usage{
iccdata1a
}
\description{
This dataset contains data from a reliability experiment where 4 judges
named J1-J4 scored 6 subjects on a continuous scale. Each subject was rated
once only.
}
\keyword{datasets}
