% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc2x3.r
\name{mse2.nointer.fn}
\alias{mse2.nointer.fn}
\title{Mean of Squares for Errors (MSE) under Models 2 & 3 without replication}
\usage{
mse2.nointer.fn(dfra)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more.  The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}
}
\description{
This function computes the MSE for both the random factorial (Model 2) and mixed factorial (Model 3) without
subject-rater Interaction. This MSE is used for calculating confidence intervals and p-values associated with the inter-rater
and intra-rater reliability coefficients.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 10, section 10.3.1,
Advanced Analytics, LLC.
}
