% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc2x3.r
\name{msi2.fn}
\alias{msi2.fn}
\title{Mean of Squares for Interaction (MSI) under ANOVA Models 2 & 3 with interaction.}
\usage{
msi2.fn(dfra)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more.  The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}
}
\description{
This function computes the MSI under both the random factorial (Model 2) and mixed factorial (Model 3) ANOVA model with
subject-rater interaction.  This MSI is used for calculating confidence intervals and p-values associated with the
inter-rater and intra-rater reliability coefficients.
coefficients under both models 2 and 3.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 9, section 9.3.1 and chapter 10, section
10.3.1. Advanced Analytics, LLC.
}
