% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc1factor.r
\name{msr1b.fn}
\alias{msr1b.fn}
\title{Mean of Squares for Raters (MSR) under ANOVA Model 1B.}
\usage{
msr1b.fn(dfra)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more.  The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}
}
\value{
This function returns a vector containing 2 values:
1. msr: the MSR. 2. r: The number of raters that participated in the experiment.
}
\description{
This MSR is associated with the one-factor ANOVA model where each rater may rate a different group of subjects. This MSR is used
for calculating confidence intervals and p-values associated with the intra-rater reliability coefficient.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 8, section 8.3.4. Advanced Analytics, LLC.
}
