% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc2x3.r
\name{msr2.fn}
\alias{msr2.fn}
\title{Mean of Squares for Raters (MSR) under ANOVA Model 2 with or without interaction.}
\usage{
msr2.fn(dfra)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more.  The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}
}
\description{
This function computes the MSR under the random factorial ANOVA model (MOdel 2). It can be used whether
or not the subject-rater interaction is assumed. The MSR is used for calculating confidence intervals and p-values associated
with the inter-rater and intra-rater reliability coefficients under model 2.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 9, section 9.3.1, Advanced Analytics, LLC.
}
