% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc1factor.r
\name{mss1a.fn}
\alias{mss1a.fn}
\title{Mean of Squares for Subjects (MSS) under ANOVA Model 1A.}
\usage{
mss1a.fn(dfra)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more.  The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}
}
\value{
This function returns a vector containing the following 2 values:
1. mss: the MSS. 2. n: The number of subjects.
}
\description{
This MSS is associated with the one-factor ANOVA model where each subject may be rated a different group of raters. This MSS is used
for calculating confidence intervals and p-values associated with the inter-rater reliability coefficient.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 8, section 8.3.1. Advanced Analytics, LLC.
}
