% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc2x3.r
\name{pvals.ICC2r.inter}
\alias{pvals.ICC2r.inter}
\title{P-values of ICC(2,1) under Model 2 with subject-rater interaction, for user-provided null values.}
\usage{
pvals.ICC2r.inter(dfra, rho.zero = 0)
}
\arguments{
\item{ratings.}{This is a data frame containing 3 columns or more. The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}

\item{rho.zero.}{This is an optional parameter that represents a vector containing an arbitrary number of null values between 0 and 1
for which a p-value will be calculated. If not specified then its default value will be 0.}
}
\value{
This function returns a vector containing p-values associated with the null values specified in the parameter rho.zero.
}
\description{
This function computes p-values for the Intraclass Correlation Coefficients (ICC) used as a measure of inter-rater reliability
under the random factorial ANOVA model (Model 2) with subject-rater interaction. The output is vector of p-values, one for each of the
null values specified in the optional rho.zero parameter, whose default value is 0.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 9, section 9.3.1 Advanced Analytics, LLC.
}
