% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc2x3.r
\name{pvals.ICC2r.nointer}
\alias{pvals.ICC2r.nointer}
\title{P-values of ICC(2,1) under Model 2 without subject-rater interaction.}
\usage{
pvals.ICC2r.nointer(dfra, rho.zero = 0)
}
\arguments{
\item{dfra}{This is a data frame containing 3 columns or more. The first column contains subject numbers (there could be duplicates
if a subject was assigned multiple ratings) and each of the remaining columns is associated with a particular rater and contains its
numeric ratings.}

\item{rho.zero}{This is an optional parameter that represents a vector containing an arbitrary number of null values between 0 and 1
for which a p-value will be calculated. If not specified then its default value will be 0.}
}
\value{
This function returns a vector containing p-values associated with the null values specified in the parameter rho.zero.
}
\description{
This function can compute several p-values associated with the Intraclass Correlation Coefficient (ICC) used to quantify inter-rater
reliability under the random factorial ANOVA model without subject-rater interaction (Model 2). This function computes
the p-value for each of the null values specified in the parameter rho.zero.
}
\references{
Gwet, K.L. (2014): \emph{Handbook of Inter-Rater Reliability - 4th ed.} chapter 9, section 9.5.1, equation 9.5.15,
Advanced Analytics, LLC.
}
